+++ runtime/docs/CHANGELOG.md
@@ -7,16 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.0.43] - 2023-06-17
+## [0.0.0.44] - 2023-06-26
 
 - Add a new ServiceConfig field to servicer config
+
+## [0.0.0.43] - 2023-06-14
+
+- Rename package import types to coreTypes for consitency and clarity
 
 ## [0.0.0.42] - 2023-06-13
 
 - Append "Hostname" to validator endpoint hostname constants
 - Promoted string literal to `RandomValidatorEndpointK8SHostname` constant
 
-## [0.0.0.41] - 2023-06-06
+## [0.0.0.41] - 2023-06-07
 
 - Adds fisherman and servicer proto configurations.
 - Renames actor hostnames